Add generic AppIcon component for rendering app icon or fallback placeholder
